backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlmodel==0.0.14
psycopg2-binary==2.9.9
alembic==1.13.1

# Google AI
google-genai>=1.0.0

# Firebase Admin
firebase-admin==6.4.0

# Utilities
python-dotenv==1.0.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data processing
pandas==2.2.0
openpyxl==3.1.2

# Testing
pytest==8.0.0
pytest-asyncio==0.21.1
httpx
